+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/ai/aiserver/AiServer.java
@@ -1,8 +1,8 @@
-package kakaotech.bootcamp.respec.specranking.domain.ai.service.aiserver;
+package kakaotech.bootcamp.respec.specranking.domain.ai.aiserver;
 
 import kakaotech.bootcamp.respec.specranking.domain.ai.dto.request.AiPostSpecRequest;
 import kakaotech.bootcamp.respec.specranking.domain.ai.dto.response.AiPostSpecResponse;
 
-public interface AiServerService {
-    AiPostSpecResponse call(AiPostSpecRequest aiPostSpecRequest);
+public interface AiServer {
+    AiPostSpecResponse analyzeSpec(AiPostSpecRequest aiPostSpecRequest);
 }
