--- src/main/java/kakaotech/bootcamp/respec/specranking/domain/ai/config/AiConfig.java
@@ -1,4 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.domain.ai.config;
-
-public class AiConfig {
-}
