+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/ai/dto/response/AiPostSpecResponse.java
@@ -4,12 +4,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@ToString
 public class AiPostSpecResponse {
 
     private String nickname;
