+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/store/service/FileStore.java
@@ -1,8 +1,9 @@
-package kakaotech.bootcamp.respec.specranking.domain.store;
+package kakaotech.bootcamp.respec.specranking.domain.store.service;
 
 import org.springframework.web.multipart.MultipartFile;
 
 public interface FileStore {
     String upload(MultipartFile multipartFile);
+
     void delete(String fileName);
 }
