+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/user/entity/User.java
@@ -1,9 +1,17 @@
 package kakaotech.bootcamp.respec.specranking.domain.user.entity;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.Table;
 import kakaotech.bootcamp.respec.specranking.domain.common.BaseTimeEntity;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.UserRole;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.UserStatus;
-import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,7 +20,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(indexes = {
-        @Index(name = "idx_nickname", columnList= "nickname")
+        @Index(name = "idx_nickname", columnList = "nickname")
 })
 public class User extends BaseTimeEntity {
 
@@ -21,8 +29,8 @@ public class User extends BaseTimeEntity {
     @Column(columnDefinition = "BIGINT UNSIGNED")
     private Long id;
 
-    @Column(name = "user_id", nullable = false, unique = true, columnDefinition = "VARCHAR(255)")
-    private String userId;
+    @Column(name = "login_id", nullable = false, unique = true, columnDefinition = "VARCHAR(255)")
+    private String loginId;
 
     @Column(nullable = false, columnDefinition = "VARCHAR(255)")
     private String password;
@@ -44,8 +52,8 @@ public class User extends BaseTimeEntity {
     @Column(nullable = false, columnDefinition = "VARCHAR(50) DEFAULT 'ACTIVE'")
     private UserStatus status;
 
-    public User(String userId, String password, String userProfileUrl, String nickname, boolean isOpenSpec) {
-        this.userId = userId;
+    public User(String loginId, String password, String userProfileUrl, String nickname, boolean isOpenSpec) {
+        this.loginId = loginId;
         this.password = password;
         this.userProfileUrl = userProfileUrl;
         this.nickname = nickname;
