--- src/main/java/kakaotech/bootcamp/respec/specranking/global/util/GetCurrentUserService.java
@@ -1,5 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.global.util;
-
-public interface GetCurrentUserService {
-    public Long getUserId();
-}
